\name{population}
\alias{population}
\docType{data}
\title{
Time Series of Population}
\description{
A time series of population from 1980 to 2001.}
\usage{data("population")}


\source{
Abbasov, A.M. and Mamedova, M.H., 2003. Application of fuzzy time series to population forecasting, Proceedings of 8th Symposion on Information Technology in Urban and Spatial Planning, Vienna University of Technology, February 25-March1, 545-552.}

\examples{
data(population)
ts.plot(population,col="red")
grid.on()
}
\keyword{datasets}
