\name{sanility}
\alias{sanility}
\docType{data}
\title{
Time Series of Crest of Sanility in Ca Mau Province
}
\description{
A time series of sanility from 2000 to 2015.}
\usage{data("sanility")}

\source{
Vo Van Tai,  Duong Ton Dam, Pham Minh Truc, Dang Kien Cuong, 2016. Forecasting crest of sanility at three main stations of Ca Mau province by fuzzy time series model.
}

\examples{
data(sanility)
ts.plot(sanility,col="red")
grid.on()
}
\keyword{datasets}
